requirements.txt
langchain==0.1.9
langchain-community==0.0.25
langchain-core==0.1.28
pypdf
python-dotenv
langchain-groq
sentence-transformers
python-docx
unstructured
chromadb
fastapi
uvicorn
tqdm
numpy
pandas
slowapi
limits
deprecated